move wcs update function to submodule
