+++ romancal/resample/resample.py
@@ -1,16 +1,11 @@
 import logging
 
 import numpy as np
-from astropy.coordinates import SkyCoord
 from roman_datamodels import datamodels, dqflags, maker_utils
-from stcal.alignment.util import (
-    compute_s_region_keyword,
-    compute_scale,
-)
 from stcal.resample import Resample
 
-from ..assign_wcs import utils
 from .exptime_resampler import ExptimeResampler
+from .l3_wcs import assign_l3_wcs
 from .resample_utils import make_output_wcs
 
 log = logging.getLogger(__name__)
@@ -218,8 +213,7 @@ class ResampleData(Resample):
                     setattr(output_model, arr_name, new_array)
 
         # assign wcs to output model
-        output_model.meta.wcs = self.output_wcs
-        gwcs_into_l3(output_model, output_model.meta.wcs)
+        assign_l3_wcs(output_model, self.output_wcs)
 
         # copy some metadata from the first input model
         # FIXME this is incorrect and these values should be
@@ -291,111 +285,3 @@ class ResampleData(Resample):
                 self.input_models.shelve(model, index, modify=False)
 
         return self.finalize()
-
-
-def gwcs_into_l3(model, wcs):
-    """Update the Level 3 wcsinfo block from a GWCS object
-
-    Parameters
-    ----------
-    model : `DataModel`
-        The model whose meta is to be updated.
-
-    wcs : `GWCS`
-        GWCS info to transfer into the `meta.wcsinfo` block
-
-    Notes
-    -----
-    Some models/parameters in the GWCS object have explicit names, such as
-    'crpix1'. However, some do not and hence have to be accessed explicitly
-    by indexing. This is fragile and will be a source of issues.
-    """
-    l3_wcsinfo = model.meta.wcsinfo
-    transform = wcs.forward_transform
-
-    l3_wcsinfo.projection = "TAN"
-    l3_wcsinfo.pixel_shape = model.shape
-
-    # Fill out image-local information
-    pixel_center = [(v - 1) / 2.0 for v in model.shape[::-1]]
-    world_center = wcs(*pixel_center)
-    l3_wcsinfo.ra_center = world_center[0]
-    l3_wcsinfo.dec_center = world_center[1]
-    l3_wcsinfo.pixel_scale_local = compute_scale(wcs, world_center)
-    l3_wcsinfo.orientat_local = calc_pa(wcs, *world_center)
-    try:
-        footprint = utils.create_footprint(wcs, model.shape)
-    except Exception as excp:
-        log.warning("Could not determine footprint due to %s", excp)
-    else:
-        l3_wcsinfo.ra_corn1 = footprint[0][0]
-        l3_wcsinfo.ra_corn2 = footprint[1][0]
-        l3_wcsinfo.ra_corn3 = footprint[2][0]
-        l3_wcsinfo.ra_corn4 = footprint[3][0]
-        l3_wcsinfo.dec_corn1 = footprint[0][1]
-        l3_wcsinfo.dec_corn2 = footprint[1][1]
-        l3_wcsinfo.dec_corn3 = footprint[2][1]
-        l3_wcsinfo.dec_corn4 = footprint[3][1]
-        l3_wcsinfo.s_region = compute_s_region_keyword(footprint)
-
-    # Fill out wcs-general information
-    try:
-        l3_wcsinfo.x_ref = -transform["crpix1"].offset.value
-        l3_wcsinfo.y_ref = -transform["crpix2"].offset.value
-    except IndexError:
-        log.warning(
-            "WCS has no clear reference pixel defined by crpix1/crpix2. Assuming reference pixel is center."
-        )
-        l3_wcsinfo.x_ref = pixel_center[0]
-        l3_wcsinfo.y_ref = pixel_center[1]
-
-    world_ref = wcs(l3_wcsinfo.x_ref, l3_wcsinfo.y_ref, with_bounding_box=False)
-    l3_wcsinfo.ra_ref = world_ref[0]
-    l3_wcsinfo.dec_ref = world_ref[1]
-
-    try:
-        cdelt1 = transform["cdelt1"].factor.value
-        cdelt2 = transform["cdelt2"].factor.value
-        l3_wcsinfo.pixel_scale = (cdelt1 + cdelt2) / 2.0
-    except IndexError:
-        l3_wcsinfo.pixel_scale = compute_scale(wcs, world_ref)
-
-    l3_wcsinfo.orientat = calc_pa(wcs, *world_ref)
-
-    try:
-        l3_wcsinfo.rotation_matrix = transform[
-            "pc_rotation_matrix"
-        ].matrix.value.tolist()
-    except Exception:
-        log.warning(
-            "WCS has no clear rotation matrix defined by pc_rotation_matrix. Calculating one."
-        )
-        rotation_matrix = utils.calc_rotation_matrix(l3_wcsinfo.orientat, 0.0)
-        l3_wcsinfo.rotation_matrix = utils.list_1d_to_2d(rotation_matrix, 2)
-
-
-def calc_pa(wcs, ra, dec):
-    """Calculate position angle at given ra,dec
-
-    Parameters
-    ----------
-    wcs : GWCS
-        The wcs in consideration.
-
-    ra, dec : float, float
-        The ra/dec in degrees.
-
-    Returns
-    -------
-    position_angle : float
-        The position angle in degrees.
-
-    """
-    delta_pix = [v for v in wcs.invert(ra, dec, with_bounding_box=False)]
-    delta_pix[1] += 1
-    delta_coord = SkyCoord(
-        *wcs(*delta_pix, with_bounding_box=False), frame="icrs", unit="deg"
-    )
-    coord = SkyCoord(ra, dec, frame="icrs", unit="deg")
-
-    return coord.position_angle(delta_coord).degree
